Remove vanilla plonk
